requirements.txt
flask>=3.0.0
python-dotenv>=1.0.1
redis>=5.0.0
pydantic>=2.7.0
langgraph>=0.2.0
mcp>=0.1.0
requests>=2.32.0
openai>=1.47.0
